Update fr.lua
Hi, you might need to sift through a bit to return to a functional code since I've added comments to explain why some things will have to leave in the final version in the CELL part. Otherwise, the translation should be OK and clean. Don't hesitate to ask if you need anything.